Move code out of the __init__ and add some extra tests.
